% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{bseqsc-package}
\alias{bseqsc}
\alias{bseqsc-package}
\title{bseqsc: Deconvolution of Bulk Sequencing Experiments using Single Cell Data}
\description{
More about what it does (maybe more than one line)
}
\author{
\strong{Maintainer}: Renaud Gaujoux \email{renaud@tx.technion.ac.il}

Authors:
\itemize{
  \item Maayan Baron \email{maayan@technion.ac.il}
}

}
