% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{bseqsc_basis}
\alias{bseqsc_basis}
\title{Building a Basis Matrix of Reference Gene Expression Profiles}
\usage{
bseqsc_basis(x, markers, clusters, samples, ct.scale = TRUE, limit = TRUE)
}
\arguments{
\item{x}{matrix of raw counts}

\item{markers}{list of cell type marker genes.
The type of gene identifiers must be the same as the ones used as feature/row names in \code{x}.}

\item{clusters}{vector giving the cell type of each cell (i.e. column) in the data \code{x}.
If \code{x} is an \code{ExpressionSet} object, it can also be the name (as a string) of the
phenotypic variable to use.}

\item{samples}{vector giving the sample of origin of each cell (i.e. column) in the data \code{x}.
If \code{x} is an \code{ExpressionSet} object, it can also be the name (as a string) of the
phenotypic variable to use.}

\item{ct.scale}{logical that indicates if the single cell expresson profiles
should be rescaled according to their cell type average total count.}

\item{limit}{logical that indicates if the returned basis matrix should only contain
cell types listed in \code{markers}.}
}
\value{
a \code{matrix} object.
}
\description{
Building a Basis Matrix of Reference Gene Expression Profiles
}
