% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{bseqsc_csdiff}
\alias{bseqsc_csdiff}
\title{Testing for Cell Type-Specific Gene Expression Differences}
\usage{
bseqsc_csdiff(formula, data = NULL, ...)
}
\arguments{
\item{formula}{csSAM model to fit specified as a formula}

\item{data}{cell type proportion matrix (samples x cell types).}

\item{...}{other arguments passed to \link[csSAM:csSAMfit]{csSAM::csSAMfit}.}
}
\description{
Uses \code{csSAM} (Shen-Orr et al. (2010)) to estimate cell type-specific
expression and test for group differences in each cell type.
}
\references{
Shen-Orr SS, Tibshirani R, Khatri P, Bodian DL, Staedtler F, Perry NM, Hastie T, Sarwal MM, Davis MM and Butte AJ (2010).
“Cell type-specific gene expression differences in complex tissues.” \emph{Nature methods}, \emph{7}(4), pp. 287-9. ISSN 1548-7105,
doi: 10.1038/nmeth.1439 (URL: http://doi.org/10.1038/nmeth.1439).
}
