% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{bseqsc_proportions}
\alias{bseqsc_proportions}
\title{Estimating Cell Type Proportions From Single Cell Data}
\usage{
bseqsc_proportions(x, reference = NULL, log = NULL, ..., verbose = TRUE)
}
\arguments{
\item{x}{Expression data from heterogenous samples}

\item{reference}{Bisis matrix of reference cell type expression profiles.
If \code{NULL} then a built-in reference for pancreatic islet cell sub-population
is used (data \link{PancreasIslet}).}

\item{log}{logical that indicates if the data \code{x} is in in log-scale.
If \code{NULL}, then log scale is inferred by \link[xbioc:is_logscale]{xbioc::is_logscale}.}

\item{...}{other arguments passed to \code{CIBERSORT}.}

\item{verbose}{logical that toggles log messages.}
}
\value{
a list with elements:
\itemize{
\item coef: matrix of estimated proportions (cell type x samples)
\item stats: statistics computed by \code{CIBERSORT}
}
}
\description{
Estimating Cell Type Proportions From Single Cell Data
}
