% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bseqscColorTheme}
\alias{bseqscColorTheme}
\title{Color Theme for BSEQsc Vignettes}
\format{An object of class \code{data.frame} with 117 rows and 2 columns.}
\usage{
bseqscColorTheme
}
\description{
Color Theme for BSEQsc Vignettes
}
\keyword{datasets}
