% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{cpm_cell_type}
\alias{cpm_cell_type}
\title{Scale Single Cell Data According to Average Cell Type Profiles}
\usage{
cpm_cell_type(x, clusters, samples)
}
\arguments{
\item{x}{matrix of raw counts}

\item{clusters}{vector giving the cell type of each cell (i.e. column) in the data \code{x}.
If \code{x} is an \code{ExpressionSet} object, it can also be the name (as a string) of the
phenotypic variable to use.}

\item{samples}{vector giving the sample of origin of each cell (i.e. column) in the data \code{x}.
If \code{x} is an \code{ExpressionSet} object, it can also be the name (as a string) of the
phenotypic variable to use.}
}
\description{
Scale Single Cell Data According to Average Cell Type Profiles
}
