% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{fitEdgeR}
\alias{fitEdgeR}
\title{Fitting EdgeR Model}
\usage{
fitEdgeR(x, formula, coef)
}
\arguments{
\item{x}{an \link[Biobase:ExpressionSet-class]{Biobase::ExpressionSet} object}

\item{formula}{EdgeR formula to fit}

\item{coef}{coefficient(s) of interest, for which a statistical test is performed.}
}
\value{
an object of class \code{TopTags}, as returned by \link[edgeR:topTags]{edgeR::topTags}.
}
\description{
Fitting EdgeR Model
}
