% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MITRIB}
\alias{MITRIB}
\title{Ribosomal and Mitochondrial Genes}
\format{An object of class \code{data.frame} with 1016 rows and 4 columns.}
\source{
http://www.genenames.org/cgi-bin/genefamilies/set/1054
}
\usage{
MITRIB
}
\description{
Ribosomal and Mitochondrial Genes
}
\keyword{datasets}
