% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PancreasIslet}
\alias{PancreasIslet}
\title{Reference Basis Matrix of Pancreatic Islet Cell Types}
\format{An object of class \code{matrix} with 61 rows and 6 columns.}
\usage{
PancreasIslet
}
\description{
Expression profiles for 6 pancreatic islet cell sub-populations,
over a set of marker genes that are characteristic of each cell type.
It is used to estimate cell type proportion from bulk tissue total gene
expression.
}
\keyword{datasets}
