% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pancreasMarkers}
\alias{pancreasMarkers}
\title{Marker Genes for Pancreas Islet Cells}
\format{An object of class \code{list} of length 6.}
\usage{
pancreasMarkers
}
\description{
List of marker genes for 6 pancreas cell sub-population:
}
\keyword{datasets}
