% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotBasis}
\alias{plotBasis}
\title{Generates Heatmap of Basis Matrix}
\usage{
plotBasis(x, markers = NULL, scale = "r1", ...)
}
\arguments{
\item{x}{matrix or cell type reference expression profiles (genes x cell types)}

\item{markers}{named list of marker genes. The llist should be named using cell type names
corresponding to column names in \code{x}.
Genes are named using the identifiers corresponding to row names in \code{x}.
Genes (rows) of \code{x} that are not in any element of \code{markers} are discarded.
Columns of \code{x} that match no element in \code{markers} are discarded.}

\item{scale}{scaling strategy. See \link[NMF:aheatmap]{NMF::aheatmap}.}

\item{...}{other arguments passed to \link[NMF:aheatmap]{NMF::aheatmap}.}
}
\description{
Draws a heatmap of reference cell type expression profiles, highlighting
marker genes.
}
