% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCellTotals}
\alias{plotCellTotals}
\title{Plots Total Count Per Cell Type}
\usage{
plotCellTotals(eset, cellType, sampleID)
}
\arguments{
\item{eset}{a matrix or \code{ExpressionSet} object.}

\item{cellType}{vector or name of phenotypic variable that contains the mapping
of each column (i.e. each cell) to a cell type.}

\item{sampleID}{vector or name of phenotypic variable that contains the mapping
of each column (i.e. each cell) to a biological sample.}
}
\value{
a \code{ggplot} object with \link[ggplot2:geom_bar]{ggplot2::geom_bar}.
}
\description{
Generate a barplot
}
\examples{
\dontshow{
options(R_CHECK_RUNNING_EXAMPLES_=TRUE) ## roxygen generated flag
}

# generate random data
x <- rscData()
# check phenotypic variables
varLabels(x)
# plot counts
plotCellTotals(x, 'cellType', 'sampleID')

}
