% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotEffectSize}
\alias{plotEffectSize}
\title{Plotting Cell Type-Specific Expression Effect Size}
\usage{
plotEffectSize(x)
}
\arguments{
\item{x}{a \code{data.frame} as returned by \link[csSAM:csTopTable]{csSAM::csTopTable}.}
}
\value{
a \code{ggplot} object.
}
\description{
Generates an horizontal barplot of cell type-specific expression
differences.
}
