% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{pvalueScatter}
\alias{pvalueScatter}
\title{P-Value Scatter Plot}
\usage{
pvalueScatter(formula, data, pval.th = 0.05, pval.th.y = pval.th,
  pval.th.x = pval.th, label.th = Inf)
}
\arguments{
\item{formula}{Formula of the form \code{yvar ~ xvar} that specify which variable in \code{data}
contains variables for the x-axis and y-axis respectively.}

\item{data}{a \code{data.frame} object containing the data to plot}

\item{pval.th}{default p-value threshold used to filter te data shown in the plot.
Only data with \eqn{p <= pval.th} are used.}

\item{pval.th.y, pval.th.x}{p-value threshold to use for the variable in the x-axis and y-axis
respectively}

\item{label.th}{-log10(p-value) threshold above which data points are labelled.
Only the points in the lower triangle are labelled.}
}
\description{
P-Value Scatter Plot
}
