% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{rscData}
\alias{rscData}
\title{Generates Random Single Cell Data}
\usage{
rscData(n = 800, p = 5000, celltypes = 15, samples = 4,
  dispersion = 0.1, .rng = NULL)
}
\arguments{
\item{n}{number of features}

\item{p}{number of cells}

\item{celltypes}{number of cell types}

\item{samples}{number of biological samples}

\item{dispersion}{dispersion parameter passed to argument \code{size} of \link[stats:rnbinom]{stats::rnbinom}.}

\item{.rng}{specification to set the random number with \link[rngtools:setRNG]{rngtools::setRNG}.
This enables to generate reproducible data.}
}
\value{
an \code{ExpressionSet} object
}
\description{
Generates random single cell data from a negative binomial distribution.
}
\examples{
\dontshow{
options(R_CHECK_RUNNING_EXAMPLES_=TRUE) ## roxygen generated flag
}

x <- rscData()
x
varLabels(x)
summary(pData(x))

}
